% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_CI.R
\name{corr.CI}
\alias{corr.CI}
\title{Correlation Coefficient with Confidence Interval}
\usage{
corr.CI(vec1, vec2)
}
\arguments{
\item{vec1}{A numeric vector.}

\item{vec2}{A numeric vector.}
}
\value{
A character string with the correlation coefficient and confidence interval.
}
\description{
Computes the correlation coefficient and a confidence interval for two vectors.
}
\examples{
vec1 <- c(1, 2, 3, 4, 5)
vec2 <- c(2, 4, 6, 8, 10)
corr.CI(vec1, vec2)
}
